+++ build_runner/test/generate/serve_test.dart
@@ -118,9 +118,8 @@ Future<ServeHandler> createHandler(List<BuildAction> buildActions,
   }));
   final actualAssets = writer.assets;
   final reader = new InMemoryRunnerAssetReader(actualAssets);
-  final packageGraph = buildPackageGraph('a', {
-    package('a', path: path.absolute('a'), includes: ['**']): []
-  });
+  final packageGraph =
+      buildPackageGraph('a', {package('a', path: path.absolute('a')): []});
   final watcherFactory = (String path) => new FakeWatcher(path);
 
   return watch_impl.watch(buildActions,